src/main/webapp/app/entities/sev-grafica-data/delete/sev-grafica-data-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { provideHttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { SevGraficaDataService } from '../service/sev-grafica-data.service';

import { SevGraficaDataDeleteDialogComponent } from './sev-grafica-data-delete-dialog.component';

describe('SevGraficaData Management Delete Component', () => {
  let comp: SevGraficaDataDeleteDialogComponent;
  let fixture: ComponentFixture<SevGraficaDataDeleteDialogComponent>;
  let service: SevGraficaDataService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [SevGraficaDataDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(SevGraficaDataDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(SevGraficaDataDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(SevGraficaDataService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
